// src/app/service/models/global.ts
export class Global{
Global: {
    NewConfirmed : number;
    TotalConfirmed: number;
    NewDeaths: number;
    TotalDeaths: number;
    NewRecovered: number;
    TotalRecovered: number;

}
Countries: {
  Country: String;
  CountryCode: string;
  NewConfirmed : number;
  TotalConfirmed: number;
  NewDeaths: number;
  TotalDeaths: number;
  NewRecovered: number;
  TotalRecovered: number;
}
Date: Date;
}